refactor: TrackDetailScreen
